+++ src/mass_driver/drivers/counter.py
@@ -1,8 +1,11 @@
-"""Simplest Patch Driver: increments a file counter"""
+"""Dummy Patch Driver: increments a file counter
+
+Simplest code we could implement that demonstrates PatchDriver capabilities.
+"""
 
 from pathlib import Path
 
-from mass_driver.patchdriver import PatchDriver, PatchOutcome, PatchResult
+from mass_driver.models.patchdriver import PatchDriver, PatchOutcome, PatchResult
 
 
 class Counter(PatchDriver):