+++ src/mass_driver/forges/__init__.py
@@ -1,4 +1,1 @@
 """All the Forges we've implemented"""
-from mass_driver.forges.github import GithubForge
-
-__all__ = ["GithubForge"]