+++ src/mass_driver/forges/github.py
@@ -2,7 +2,7 @@
 
 from github import Github
 
-from mass_driver.forge import BranchName, Forge
+from mass_driver.models.forge import BranchName, Forge
 
 
 class GithubForge(Forge):
@@ -32,7 +32,7 @@ class GithubForge(Forge):
         return pr.html_url
 
     def get_pr(self, forge_repo: str, pr_id: str):
-        """Send a PR with msg on upstream of repo at repo_path, for given branch"""
+        """Get the PR by ID on forge_repo"""
         api = Github(self.auth_token)
         repo = api.get_repo(forge_repo)
         return repo.get_pull(int(pr_id))